requirements.txt
mysql-connector-python
nextcord
